Flexible Transfer Model (FTM) Linear Model Class

An S4 class designed to encapsulate the necessary components for a flexible transfer model linear model. This class stores matrices that represent the weighted cross-product of predictors and the cross-product of predictors with outcomes, essential for the model calculations.

Usage

This class is internally used by functions within the package to perform regression analysis. It is not intended for direct use.

Slots

XtX

matrix representing the cross product of the predictor variables.

Xty

matrix representing the cross product of the predictor variables and outcomes.

s

numeric representing the optimal Ridge penalty from cross-validation (if available).